#438 r4 rc=1 FIX: torch.linalg.inv has no BFloat16 kernel — under the run-global CUDA bf16 autocast the pose-carrier homography chain fed it bf16 H (H100 rc=1 @415.4s, ~$0.58; compile fix CONFIRMED working — Triton autotune ran). forward_with_xi now runs the ENTIRE warp fp32 with autocast disabled (geometry is precision-critical: bf16 quantizes ~1164px projective coords to multi-pixel error; matches the fp32 NumPy/MLX reference warp). Sister sweep: jacobian probe .float() before .numpy() (numpy has no bf16; would crash at first probe epoch); dead numpy import removed; +1 REAL bf16-autocast test ($0 class). pointer UNMOVED (apparatus)

+++ src/tac/cuda_v9_controller_runtime.py
@@ -12,7 +12,6 @@ from dataclasses import dataclass
 import copy
 from typing import Any
 
-import numpy as np
 
 from tac.witness_control.birth_completion import (
     BirthCompletionController,
@@ -758,7 +757,10 @@ def torch_pose_jacobian_conditioning(
             jacobian = torch.autograd.functional.jacobian(
                 fn, xi0, create_graph=False, vectorize=True
             )
-            conds.append(conditioning(jacobian.detach().cpu().numpy()))
+            # .float() before .numpy(): numpy has no bfloat16, and under the
+            # run-global CUDA autocast the pose forward can emit bf16 (sister
+            # of the r4 linalg.inv bf16 crash class; $0-testable).
+            conds.append(conditioning(jacobian.detach().float().cpu().numpy()))
     aggregate = aggregate_conditioning(conds, sigma_floor=float(sigma_floor))
     return {**aggregate, "pair_indices": idxs}
